backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
twilio==9.3.0
python-dotenv==1.0.1